+++ system/helpers/captcha_helper.php
@@ -43,15 +43,31 @@ if ( ! function_exists('create_captcha'))
 	/**
 	 * Create CAPTCHA
 	 *
-	 * @param	array	array of data for the CAPTCHA
-	 * @param	string	path to create the image in
-	 * @param	string	URL to the CAPTCHA image folder
-	 * @param	string	server path to font
+	 * @param	array	$data		data for the CAPTCHA
+	 * @param	string	$img_path	path to create the image in
+	 * @param	string	$img_url	URL to the CAPTCHA image folder
+	 * @param	string	$font_path	server path to font
 	 * @return	string
 	 */
 	function create_captcha($data = '', $img_path = '', $img_url = '', $font_path = '')
 	{
-		$defaults = array('word' => '', 'img_path' => '', 'img_url' => '', 'img_width' => '150', 'img_height' => '30', 'font_path' => '', 'expiration' => 7200, 'word_length' => 8, 'pool' => '0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ');
+		$defaults = array(
+			'word'		=> '',
+			'img_path'	=> '',
+			'img_url'	=> '',
+			'img_width'	=> '150',
+			'img_height'	=> '30',
+			'font_path'	=> '',
+			'expiration'	=> 7200,
+			'word_length'	=> 8,
+			'pool'		=> '0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ',
+			'colors'	=> array(
+				'background'	=> array(255,255,255),
+				'border'	=> array(153,102,102),
+				'text'		=> array(204,153,153),
+				'grid'		=> array(255,182,182)
+			)
+		);
 
 		foreach ($defaults as $key => $val)
 		{
@@ -122,15 +138,19 @@ if ( ! function_exists('create_captcha'))
 
 		// -----------------------------------
 		//  Assign colors
-		// -----------------------------------
-		$bg_color	= imagecolorallocate($im, 255, 255, 255);
-		$border_color	= imagecolorallocate($im, 153, 102, 102);
-		$text_color	= imagecolorallocate($im, 204, 153, 153);
-		$grid_color	= imagecolorallocate($im, 255, 182, 182);
-		$shadow_color	= imagecolorallocate($im, 255, 240, 240);
+		// ----------------------------------
 
-		//  Create the rectangle
-		ImageFilledRectangle($im, 0, 0, $img_width, $img_height, $bg_color);
+		is_array($colors) OR $colors = $defaults['colors'];
+
+		foreach (array_keys($default['colors']) as $key)
+		{
+			// Check for a possible missing value
+			is_array($colors[$key]) OR $colors[$key] = $defaults['colors'][$key];
+			$colors[$key] = imagecolorallocate($im, $colors[$key][0], $colors[$key][1], $colors[$key][2]);
+		}
+
+		// Create the rectangle
+		ImageFilledRectangle($im, 0, 0, $img_width, $img_height, $colors['background']);
 
 		// -----------------------------------
 		//  Create the spiral pattern
@@ -151,7 +171,7 @@ if ( ! function_exists('create_captcha'))
 			$rad1 = $radius * (($i + 1) / $points);
 			$x1 = ($rad1 * cos($theta)) + $x_axis;
 			$y1 = ($rad1 * sin($theta)) + $y_axis;
-			imageline($im, $x, $y, $x1, $y1, $grid_color);
+			imageline($im, $x, $y, $x1, $y1, $colors['grid']);
 			$theta -= $thetac;
 		}
 
@@ -178,19 +198,19 @@ if ( ! function_exists('create_captcha'))
 			if ($use_font === FALSE)
 			{
 				$y = rand(0 , $img_height / 2);
-				imagestring($im, $font_size, $x, $y, $word[$i], $text_color);
+				imagestring($im, $font_size, $x, $y, $word[$i], $colors['text']);
 				$x += ($font_size * 2);
 			}
 			else
 			{
 				$y = rand($img_height / 2, $img_height - 3);
-				imagettftext($im, $font_size, $angle, $x, $y, $text_color, $font_path, $word[$i]);
+				imagettftext($im, $font_size, $angle, $x, $y, $colors['text'], $font_path, $word[$i]);
 				$x += $font_size;
 			}
 		}
 
 		// Create the border
-		imagerectangle($im, 0, 0, $img_width - 1, $img_height - 1, $border_color);
+		imagerectangle($im, 0, 0, $img_width - 1, $img_height - 1, $colors['border']);
 
 		// -----------------------------------
 		//  Generate the image